+++ .github/workflows/linux.yml
@@ -92,7 +92,7 @@ jobs:
                  exit 1
           fi
           
-          go test -timeout 20m -v -race -cover -tags=debug -failfast -coverpkg=$(cat pkgs.txt) -coverprofile=./coverage-ci/kafka.out -covermode=atomic jobs_kafka_test.go
+          go test -timeout 20m -v -race -cover -tags=debug -failfast -coverpkg=$(cat pkgs.txt) -coverprofile=./coverage-ci/kafka.out -covermode=atomic jobs_kafka_test.go jobs_kafka_tls_test.go
           
           docker compose -f env/docker-compose-kafka.yaml down
