+++ tests/jobs_kafka_test.go
@@ -1,4 +1,4 @@
-package durability
+package tests
 
 import (
 	"fmt"
@@ -860,93 +860,6 @@ func TestKafkaPingOk(t *testing.T) {
 	assert.Equal(t, 1, oLogger.FilterMessage("ping kafka: ok").Len())
 }
 
-func TestKafkaTLS(t *testing.T) {
-	t.Run("TestKafkaTLS", pushJobTest("configs/.rr-kafka-tls.yaml", "127.0.0.1:6002", "test-1-tls"))
-}
-
-func TestKafkaTLSRootCA(t *testing.T) {
-	t.Run("TestKafkaTLSRootCA", pushJobTest("configs/.rr-kafka-tls-root-ca.yaml", "127.0.0.1:6003", "test-1-tls-root-ca"))
-}
-
-func pushJobTest(env string, rpcAddress string, pipeline string) func(t *testing.T) {
-	return func(t *testing.T) {
-		cont := endure.New(slog.LevelDebug)
-
-		cfg := &config.Plugin{
-			Version: "v2023.1.0",
-			Path:    env,
-			Prefix:  "rr",
-		}
-
-		l, oLogger := mocklogger.ZapTestLogger(zap.DebugLevel)
-		err := cont.RegisterAll(
-			cfg,
-			&server.Plugin{},
-			&rpcPlugin.Plugin{},
-			&jobs.Plugin{},
-			&kp.Plugin{},
-			l,
-			&resetter.Plugin{},
-			&informer.Plugin{},
-		)
-		assert.NoError(t, err)
-
-		err = cont.Init()
-		require.NoError(t, err)
-
-		ch, err := cont.Serve()
-		require.NoError(t, err)
-
-		sig := make(chan os.Signal, 1)
-		signal.Notify(sig, os.Interrupt, syscall.SIGINT, syscall.SIGTERM)
-
-		wg := &sync.WaitGroup{}
-		wg.Add(1)
-
-		stopCh := make(chan struct{}, 1)
-
-		go func() {
-			defer wg.Done()
-			for {
-				select {
-				case e := <-ch:
-					assert.Fail(t, "error", e.Error.Error())
-					err = cont.Stop()
-					if err != nil {
-						assert.FailNow(t, "error", err.Error())
-					}
-				case <-sig:
-					err = cont.Stop()
-					if err != nil {
-						assert.FailNow(t, "error", err.Error())
-					}
-					return
-				case <-stopCh:
-					// timeout
-					err = cont.Stop()
-					if err != nil {
-						assert.FailNow(t, "error", err.Error())
-					}
-					return
-				}
-			}
-		}()
-
-		time.Sleep(time.Second * 3)
-		t.Run("PushPipeline", helpers.PushToPipe(pipeline, false, rpcAddress))
-		time.Sleep(time.Second * 5)
-		t.Run("DestroyPipelines", helpers.DestroyPipelines(rpcAddress, pipeline))
-		time.Sleep(time.Second * 5)
-
-		stopCh <- struct{}{}
-		wg.Wait()
-
-		assert.Equal(t, 1, oLogger.FilterMessageSnippet("job was pushed successfully").Len())
-		assert.Equal(t, 1, oLogger.FilterMessageSnippet("job processing was started").Len())
-		assert.Equal(t, 1, oLogger.FilterMessageSnippet("job was processed successfully").Len())
-	}
-}
-
 func declarePipe(topic string) func(t *testing.T) {
 	return func(t *testing.T) {
 		conn, err := net.Dial("tcp", "127.0.0.1:6001")
@@ -963,10 +876,10 @@ func declarePipe(topic string) func(t *testing.T) {
 				"auto_create_topics_enable": "true",
 
 				"producer_options": `{
-                     "max_message_bytes": 1000,
-                     "required_acks": "LeaderAck",
-                     "compression_codec": "snappy",
-				     "disable_idempotent": true
+					"max_message_bytes": 1000,
+					"required_acks": "LeaderAck",
+					"compression_codec": "snappy",
+					"disable_idempotent": true
 				}`,
 
 				"consumer_options": consumer,
@@ -995,15 +908,15 @@ func declarePipeCG(topic string) func(t *testing.T) {
 				"auto_create_topics_enable": "true",
 
 				"group_options": `{
-                     "group_id": "foo-bar",
-				     "block_rebalance_on_poll": true
+					"group_id": "foo-bar",
+					"block_rebalance_on_poll": true
 				}`,
 
 				"producer_options": `{
-                     "max_message_bytes": 1000,
-                     "required_acks": "LeaderAck",
-                     "compression_codec": "snappy",
-				     "disable_idempotent": true
+					"max_message_bytes": 1000,
+					"required_acks": "LeaderAck",
+					"compression_codec": "snappy",
+					"disable_idempotent": true
 				}`,
 
 				"consumer_options": consumer,
